Added fixed blimp IDs
Restart program now takes in the blimps ID. Each pi has to start the program WITH its ID as a parameter.
- ssh into pi
- cd /etc
- sudo nano rc.local
- add that pi's ID as a parameter to the program

+++ blimpVision/restartProgram.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
-if [ "$1" != "" ]; then
+if [ "$1" != "" ] && [ "$2" != "" ]; then
 	hostname=$1
+	blimpID=$2
 	echo "Killing program."
 	ssh pi@$hostname 'sudo killall -9 piStereoVision'
 	echo "Waiting 6 seconds before restarting..."
 	sleep 6
 	echo "Starting program."
-	ssh pi@$hostname './piStereoVision/build/piStereoVision'
+	ssh pi@$hostname "./piStereoVision/build/piStereoVision $blimpID"
 else
-	echo "Usage: restartProgram.sh [hostname]"
+	echo "Usage: restartProgram.sh [hostname] [blimpID]"
 fi
